refactor(flterm): move selection ownership into input
