+++ packages/flterm/lib/src/controller/terminal_controller.dart
@@ -10,8 +10,8 @@ import '../foundation.dart';
 import '../input/input_encoder.dart';
 import '../input/input_message.dart';
 import '../input/input_modifiers.dart';
+import '../input/selection_session.dart';
 import '../input/text_input_session.dart';
-import '../interaction/selection_session.dart';
 import '../links/link_interaction.dart';
 import '../links/link_settings.dart';
 import '../view/compression_scheduler.dart';
